+++ FriedVM/Types/BaseValue.h
@@ -2,23 +2,7 @@
 #include <cstdint>
 #include <memory>
 #include "../Logger.h"
-//#include "Types.h"
-
-enum ValueType
-{
-    vt_raw = 0,
-    vt_uint8_t = 1,
-    vt_uint16_t = 2,
-    vt_uint32_t = 3,
-    vt_uint64_t = 4,
-    vt_int8_t = 5,
-    vt_int16_t = 6,
-    vt_int32_t = 7,
-    vt_int64_t = 8,
-    vt_float_t = 9,
-    vt_double_t = 10,
-    vt_pointer = 11,
-};
+#include "Types.h"
 
 struct BaseValue
 {
@@ -47,23 +31,7 @@ struct BaseValue
     }
 
     // Function to get the size of the element based on the type
-    constexpr static size_t getTypeSize(ValueType type) {
-        switch (type) {
-        case vt_raw: return 4;
-        case vt_uint8_t: return sizeof(uint8_t);
-        case vt_uint16_t: return sizeof(uint16_t);
-        case vt_uint32_t: return sizeof(uint32_t);
-        case vt_uint64_t: return sizeof(uint64_t);
-        case vt_int8_t: return sizeof(int8_t);
-        case vt_int16_t: return sizeof(int16_t);
-        case vt_int32_t: return sizeof(int32_t);
-        case vt_int64_t: return sizeof(int64_t);
-        case vt_float_t: return sizeof(float);
-        case vt_double_t: return sizeof(double);
-        case vt_pointer: return 4;
-        default: return 0;
-        }
-    }
+    constexpr static size_t getTypeSize(ValueType type);
 
     // Static factory methods for creating BaseValue instances
     static BaseValue makeRaw(uint8_t* data, size_t len) {
@@ -94,14 +62,8 @@ struct BaseValue
         type_index = type;
     }
 
-    bool isNumber() const {
-        // Return true if the type is one of the numeric types
-        return type_index == vt_uint8_t || type_index == vt_uint16_t ||
-            type_index == vt_uint32_t || type_index == vt_uint64_t ||
-            type_index == vt_int8_t || type_index == vt_int16_t ||
-            type_index == vt_int32_t || type_index == vt_int64_t ||
-            type_index == vt_float_t || type_index == vt_double_t;
-    }
+    bool isNumber() const;
+    std::string toString() const;
 
     //// Cast function to convert BaseValue to a different type
     //BaseValue Cast(ValueType targetType) const {
@@ -166,37 +128,5 @@ struct BaseValue
         return result;
     }
 
-    // The main Add method to dispatch based on type
-    BaseValue Add(const BaseValue& v1, const BaseValue& v2) {
-        // Validate if both values are numbers
-        if (!v1.isNumber() || !v2.isNumber()) {
-            throw std::runtime_error("Cannot add non-number values");
-        }
-        //if (v1.type_index != v2.type_index || v1.length != v2.length) {
-        //    // Cast v2 to the same type as v1
-        //    BaseValue casted = v2.Cast(v1.type_index);
-        //    return Add(v1, casted);  // Now they are both of the same type
-        //}
-
-        BaseValue result(v1.type_index, v1.length);
-
-        // Dispatch based on type
-        switch (v1.type_index) {
-        case vt_uint8_t:   result = AddTyped<uint8_t>(v1, v2); break;
-        case vt_uint16_t:  result = AddTyped<uint16_t>(v1, v2); break;
-        case vt_uint32_t:  result = AddTyped<uint32_t>(v1, v2); break;
-        case vt_uint64_t:  result = AddTyped<uint64_t>(v1, v2); break;
-        case vt_int8_t:    result = AddTyped<int8_t>(v1, v2); break;
-        case vt_int16_t:   result = AddTyped<int16_t>(v1, v2); break;
-        case vt_int32_t:   result = AddTyped<int32_t>(v1, v2); break;
-        case vt_int64_t:   result = AddTyped<int64_t>(v1, v2); break;
-        case vt_float_t:   result = AddTyped<float>(v1, v2); break;
-        case vt_double_t:  result = AddTyped<double>(v1, v2); break;
-        default:
-            throw std::runtime_error("Unsupported type in Add");
-        }
-
-        return result;
-    }
-
+    BaseValue Add(const BaseValue& v1, const BaseValue& v2);
 };
