+++ FriedVM/Types/Types.h
@@ -1,0 +1,17 @@
+#pragma once
+
+enum ValueType
+{
+    vt_raw = 0,
+    vt_uint8_t = 1,
+    vt_uint16_t = 2,
+    vt_uint32_t = 3,
+    vt_uint64_t = 4,
+    vt_int8_t = 5,
+    vt_int16_t = 6,
+    vt_int32_t = 7,
+    vt_int64_t = 8,
+    vt_float_t = 9,
+    vt_double_t = 10,
+    vt_pointer = 11,
+};
